requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
transformers>=4.40.0
grad-cam>=1.5.0
scikit-learn>=1.3.0
pandas>=2.0.0
Pillow>=10.0.0
matplotlib>=3.8.0
tqdm
numpy
kagglehub>=1.0.0
